requirements.txt
pyautogui>=0.9.54
opencv-python>=4.8.0.74
Pillow>=10.1.0
pynput>=1.7.6
PyInstaller>=5.13.0
